Hackathon/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.0.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.40
requests>=2.31.0
scikit-learn>=1.3.0
plotly>=5.0.0
xgboost>=2.1.0
catboost>=1.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
google-generativeai>=0.8.0
